refactor(web): extract render helpers and approval styles to meet 350-LOC ceiling
