+++ src/web-server/frontend/index.html
@@ -21,6 +21,7 @@
       href="data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 32 32'%3E%3Ccircle cx='16' cy='16' r='10' fill='%23f9854b'/%3E%3C/svg%3E"
     />
     <link rel="stylesheet" href="/styles.css" />
+    <link rel="stylesheet" href="/approvals.css" />
   </head>
   <body>
     <header class="topbar">
